vectorize the prediction of neighbor pixels from idct data

+++ src/lepton/idct.hh
@@ -1,2 +1,2 @@
 class AlignedBlock;
-void idct(const AlignedBlock &block, const uint16_t quantization[64], int32_t outp[64], bool ignore_dc);
+void idct(const AlignedBlock &block, const uint16_t quantization[64], int16_t outp[64], bool ignore_dc);
